Nested source codes to allow e2e samples and compatibility tests without sharing node_modules
To make testing compatibility and usage with clean install of Emmett modules it's needed to have structure that won't share node_modules.
This commits nests all the current source codes under src. Thanks to that it'll be possible to add samples that are not falling back to the source codes instead of using bundled Emmett code from node_modules.
That will allow catching bugs that were caused by transpilation, bundling, etc.