Made event store features tests internal not to expose node packages in bundled Emmett code
The issue is that `esbuild` (for `commonjs`) stripes the 'node:' prefix for targets different from the node (or using node earlier than 14) as they cannot handle packages with prefix like `node:test`. See more in:
https://github.com/evanw/esbuild/commit/6b4f97086c6247434ba95031c1e4335ef67d10b6

`Tsup` that I'm using for bundling also seems not to be handling that fully as it's built on top of `esbuild`: https://github.com/egoist/tsup/issues/1003

As there's no dependency to `test` package, then bundled code fails to load in commonjs environment. One solution is to add dependency to shim, as @alex-laycalvert suggested in #60. But I'm reluctant to do it, as I don't want to target older node environments than Node LTS version (currently >= 20.11.1). I also don't want to add dependency on the unmaintained shim package with potential vulnerabilities.

I could try to change tsconfig target from ESNext to Node, but at the moment I don't want to have this limitation, as I'd like to have the main Emmett package working both with backend and frontend.

The change removes the event store matrix tests, as they don't need to be exposed now. They're still available as js map files to make internal referencing work correctly.

This change should be seen more as a fix rather than solving the root cause. I need to think longer on alternative solutions like having internal project for testing event stores and not having this test at all.

Fixes #59,
Closes #60

Besides that, following the boy scout rule I removed the `no-floating-promise` ingores as I found the way of handling them using `void` keyword before describes and it. And that Fixes #16

@alex-laycalvert @thiagomini @mkubasz FYI