Added PostgreSQL event store consumer
Consumer will be responsible for pulling messages and feeding it to subscriptions.

A single consumer can be feeding multiple subscriptions. Thanks to that user can choose the subscriptions isolation. Havina a single consumer for multiple subscriptions makes it less chatty, as there'll be only a single process pulling. Multiple consumers can be spin up to scale subscriptions separately or handle rebuilds.

Now, the same PostgreSQLSubscription can be feeded from other sources than event store, but storing checkpointing in the same place, and handling projections (so e.g. connecting Kafka, or logical replication instead of the message pulling).

Message Puller was taken out to a dedicated file.