Refactored Inline Projection
Now inline read models do not have redundant nesting and split into "short" and "details".

The append checks if there's a need to apply projections. If yes, then loads the projections (and expected version). Then projections are used to build the update.

Aligned also typing to be closer to the other projecitons.

Removed async projections for now, as they're a bit more complex and can come in a dedicated PR.