[MongoDB Event Store] Projection Registration & Helper for single stream projections (#127)
* updated projection registration and tests

* updated comments

* removed custom bigint serializer for useBigInt64 native mongodb option

* remove JS slice and include slice for mongoDB level

* update test to add in checking for slice of events

* fixed tests and updated projection handling

* remove reading from stream when appending

* update tests add to/from options

* avoid reading stream again on append

* Refactored Inline Projection

Now inline read models do not have redundant nesting and split into "short" and "details".

The append checks if there's a need to apply projections. If yes, then loads the projections (and expected version). Then projections are used to build the update.

Aligned also typing to be closer to the other projecitons.

Removed async projections for now, as they're a bit more complex and can come in a dedicated PR.

* Extended MongoDB read model metadata with schema version

This will be useful for rebuilding projections read models when new version was added

---------

Co-authored-by: Oskar Dudycz <oskar.dudycz@gmail.com>