Made the MongoDB event store explicit on whether it can close MongoDB client or not
So far MongoDB event store was always exposing close method. That's redundant for the case when client is passed through options.

Now it'll only be exposed if store is created based on the connection string.

Following the boy scout rule, I added also a general Closeable interface to make easier uniform handling such cases in the future.