Added default event type to consumers
Thanks to that processors will be automatically typed by that. This can be useful if we have multiple processors for the same stream subscription.

It's optional and falls back to event, so each processor can still have it's own event type (e.g. when subscribing to all events).

Other option is to put a union of all events in consumer and then use just a subset of those types in processors.