Added order by in readStream to ensure messages ordering
Based on the @dawidranda finding in #239, it seems that sometimes
PostgreSQL event store is not returning events in the proper order.

Emmett has an index on stream_id and stream position which is ascending,
so when reading it's using the index and returning messages in right
order.

If it doesn't work for you then it probably means that for some reason
Postgres is not using index. The potential reason is that maybe it only
has rows from a single stream, and then it decides that doing full scan
will be faster than using index.

I wasn't able to reproduce on my local box the issues, even when adding
tests simulating scenario described in #239, but adding ORDER BY won't
do harm, as it should still use index, so we should not get the
performance penalty.

Fixes #239