Fixed EventStoreDB category stream checkpointing
Refactored checkpointing handling to allow more advanced checkpointing
strategies. EventStoreDB for category stream should be using the
category stream position which is different from the original stream
position that was used (and that's mapped to the stream) and global
position.

Now consumer fills checkpoint property in metadata instead of passing
custom function to distinguish the checkpointing detection. Now
checkpointer can just use this property. This will be also useful in the
future for e.g. Kafka consumer, as it will use custom offset.