Removed dependency on web-streams-pollyfil and removed emmett-shims package
The initial idea to use streaming shims was to give a streaming support
to event store both on web and node.js, still, this pollyfil dependency
can cause issues in compatibility with other tools, like Svelte.

I also have now other idea how to approach that, probably with a
dedicated package. I outlined it here:
https://www.architecture-weekly.com/p/compilation-isnt-just-for-programming

For now moved streaming code to emmett-postgresql not to lose it, but
it'll be removed from it and moved to a dedicated package eventually.

Removed also stream function from InMemoryEventStore, which I think/hope
noone used